Incorporation of Andy's initial develop
Incorporates the various development Andy had done on the osm2ma
conversion tools prior to the move to GitHub.
