+++ osm2ma/fixtures.py
@@ -1,0 +1,19 @@
+#-------------------------------------------------------------------------------
+# Name:        module1
+# Purpose:
+#
+# Author:      asmith
+#
+# Created:     17/09/2014
+# Copyright:   (c) asmith 2014
+# Licence:     <your licence>
+#-------------------------------------------------------------------------------
+
+rawconf_good =
+
+rawconf_wrong_schema =
+
+rawconf_invalid_heirarchy =
+
+xWalk_good =
+
