+++ osm2ma/test_configengine.py
@@ -1,0 +1,67 @@
+#This file was originally generated by PyScripter's unitest wizard
+
+import unittest
+import configengine
+
+class TestRowsFromSheet(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test__iter__(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def testnext(self):
+        pass
+
+class TestConfigXWalk(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test_init_db_tables(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test_populate_config_table(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test_populate_scratch_table(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test_populate_shpfile_table(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test_init_db_funcs(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def testgetXWalk(self):
+        pass
+
+class TestGlobalFunctions(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def testxWalkFromRawConfig(self):
+        pass
+
+if __name__ == '__main__':
+    unittest.main()
