+++ osm2ma/test_raw_conf_loader.py
@@ -1,0 +1,66 @@
+#This file was originally generated by PyScripter's unitest wizard
+
+import unittest
+import osm2ma.raw_config_loader
+
+class TestExcelLoader(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    def testget_raw_config(self):
+        pass
+
+class TestRawConfig(unittest.TestCase):
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    def test_isColumnNamesValid(self):
+        pass
+
+    def test_isColumnCountValid(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test_isRowCountValid(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def getRawConfigIterator(self):
+        pass
+
+class TestRawConfigIterator(unittest.TestCase):
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def test__iter__(self):
+        pass
+
+    @unittest.skip("not implenemted")
+    def testnext(self):
+        pass
+
+
+class TestGlobalFunctions(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    def testrawConfigFromFile(self):
+        pass
+
+if __name__ == '__main__':
+    unittest.main()
