--- osm2shp-wrapper/test.psproj
@@ -1,45 +1,0 @@
-﻿[PyScripter]
-Version=2.5.3.0
-
-[Project]
-ClassName=TProjectRootNode
-StoreRelativePaths=TRUE
-ShowFileExtensions=FALSE
-
-[Project\ChildNodes\Node0]
-ClassName=TProjectFilesNode
-
-[Project\ChildNodes\Node1]
-ClassName=TProjectRunConfiguationsNode
-
-[Project\ChildNodes\Node1\ChildNodes\Node0]
-ClassName=TProjectRunConfiguationNode
-Name=unit-tests
-
-[Project\ChildNodes\Node1\ChildNodes\Node0\RunConfig]
-ScriptName=D:\work\custom-software-group\osm-ma-shp\maosmtoshape_unittests.py
-EngineType=peRemote
-ReinitializeBeforeRun=TRUE
-WorkingDir=$[ActiveScript-Dir]
-WriteOutputToFile=FALSE
-OutputFileName=$[ActiveScript-NoExt].log
-AppendToFile=FALSE
-
-[Project\ChildNodes\Node1\ChildNodes\Node0\RunConfig\ExternalRun]
-Caption=External Run
-Description=Run script using an external Python Interpreter
-ApplicationName=C:\OSGeo4W64\OSGeo4W.bat
-Parameters=python.exe -i $[ActiveScript-Short]
-WorkingDirectory=$[ActiveScript-Dir]
-ShortCut=0
-MessagesFormat=$[FileName] $[LineNumber]
-
-[Project\ChildNodes\Node1\ChildNodes]
-Count=1
-
-[Project\ChildNodes]
-Count=2
-
-[Project\ExtraPythonPath]
-Count=0
-
